+++ test/functional/mongodb/basic/embedded-columns/entity/EditHistory.ts
@@ -1,0 +1,11 @@
+import {Column} from "../../../../../../src/decorator/columns/Column";
+
+export class EditHistory {
+
+    @Column()
+    title: string;
+
+    @Column()
+    text: string;
+
+}