+++ test/functional/mongodb/basic/embedded-columns/entity/ExtraInformation.ts
@@ -1,0 +1,9 @@
+import {Column} from "../../../../../../src/decorator/columns/Column";
+import {EditHistory} from "./EditHistory";
+
+export class ExtraInformation {
+
+    @Column(type => EditHistory)
+    lastEdit: EditHistory;
+
+}